Add FileResponse for asynchronous file return
A shared pointer to a FileResponse can be returned from a PathHandler's handle()
method to asynchronously return the indicated file back to a client.
It understands and supports the Range header for non-multipart byte ranges along
with the If-Range, If-Modified-Since, and If-Unmodified-Since conditional headers.

Additionally, if enabled during construction and the client supports it,
the file data being sent back to the client is compressed using the
built in zlib support for deflate Content-Encoding.
This does not compress ahead of time and cache the compressed data,
instead the data is compressed while it is sent back to the client.
Because of this, content length is not known ahead of time and resume
features are disabled.